+++ app/oryx-app-mllib/src/test/java/com/cloudera/oryx/app/batch/mllib/kmeans/KMeansHyperParamTuningIT.java
@@ -79,7 +79,7 @@ public final class KMeansHyperParamTuningIT extends AbstractKMeansIT {
     ClusteringModel clusteringModel = (ClusteringModel) rootModel;
 
     // Should have picked highest k
-    assertEquals(100, clusteringModel.getNumberOfClusters());
+    assertEquals(100, clusteringModel.getNumberOfClusters().intValue());
   }
 
 }
